Join the various definitions together, and fix the way they are exported
