--- definitions/db/data_types.d.ts
@@ -1,7 +1,0 @@
-export interface IDataTypes {
-    [type: string]: {
-        convert: (v: any) => any;
-    };
-}
-declare const dataTypes: IDataTypes;
-export default dataTypes;
