--- definitions/index.d.ts
@@ -1,43 +1,0 @@
-import * as API from './api';
-import Application from './application';
-import Composer from './composer';
-import Controller from './controller';
-import Database from './db/database';
-import Daemon from './daemon';
-import GraphQuery from './graph_query';
-import ItemArray from './item_array';
-import Migration from './db/migration';
-import Model from './model';
-import ModelArray from './model_array';
-import ModelFactory from './model_factory';
-import RelationshipGraph from './relationship_graph';
-import Router from './router';
-import Scheduler from './scheduler';
-import SchemaGenerator from './db/schema_generator';
-declare var _default: {
-    API: typeof API;
-    APIResource: any;
-    Application: typeof Application;
-    Controller: typeof Controller;
-    Composer: typeof Composer;
-    CLI: any;
-    Daemon: typeof Daemon;
-    Database: typeof Database;
-    GraphQuery: typeof GraphQuery;
-    ItemArray: typeof ItemArray;
-    Migration: typeof Migration;
-    Mime: any;
-    Model: typeof Model;
-    ModelArray: typeof ModelArray;
-    ModelFactory: typeof ModelFactory;
-    RelationshipGraph: typeof RelationshipGraph;
-    Router: typeof Router;
-    Scheduler: typeof Scheduler;
-    SchemaGenerator: typeof SchemaGenerator;
-    my: {
-        Config?: any;
-        Schema?: any;
-        bootstrapper?: any;
-    };
-};
-export default _default;
