--- definitions/router.d.ts
@@ -1,4 +1,0 @@
-import * as fxn from 'fxn';
-declare class Router extends fxn.Router {
-}
-export default Router;
