--- definitions/scheduler.d.ts
@@ -1,8 +1,0 @@
-import fxn from 'fxn';
-/**
- * Use to delegate tasks minutely, hourly, daily, or weekly.
- * @class
- */
-declare class Scheduler extends fxn.Scheduler {
-}
-export default Scheduler;
