--- definitions/utilities.d.ts
@@ -1,2 +1,0 @@
-declare var _default: any;
-export default _default;
